Putting MSE calculation into for loop
Adding in a for loop to compare a model from the training set with seismic point depths from the testing set, scoring the points to get a MSE estimation. By putting it into a loop eventually a histogram can be created of all MSE's to find how good the model is without the seismic estimates.